requirements.txt
astroid==2.3.3
autopep8==1.4.4
beautifulsoup4==4.8.1
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.8
django-autocomplete-light==3.5.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.0.1
django-ckeditor==5.8.0
django-contrib-comments==1.9.2
django-extensions==2.2.5
django-js-asset==1.2.2
django-mptt==0.10.0
django-rosetta==0.9.3
django-settings-export==1.2.1
django-tagging==0.4.6
django-xmlrpc==0.1.8
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mots-vides==2015.5.11
pep8==1.7.1
Pillow==6.2.1
polib==1.1.0
pycodestyle==2.5.0
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pyparsing==2.4.5
pytz==2019.3
regex==2019.11.1
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
whitenoise==4.1.4
wrapt==1.11.2
